refactor: remove double in favor of float

+++ src/main/java/com/rinha/frauddetector/adapter/engine/Distance.java
@@ -2,5 +2,6 @@ package com.rinha.frauddetector.adapter.engine;
 
 @FunctionalInterface
 public interface Distance<T> {
-  double calculate(T a, T b);
+  float calculate(T a, T b);
+
 }
