+++ src/main/java/com/rinha/frauddetector/domain/NormalizationConstants.java
@@ -1,10 +1,10 @@
 package com.rinha.frauddetector.domain;
 
 public record NormalizationConstants(
-    double max_amount,
-    double max_installments,
-    double amount_vs_avg_ratio,
-    double max_minutes,
-    double max_km,
-    double max_tx_count_24h,
-    double max_merchant_avg_amount) {}
+    float max_amount,
+    float max_installments,
+    float amount_vs_avg_ratio,
+    float max_minutes,
+    float max_km,
+    float max_tx_count_24h,
+    float max_merchant_avg_amount) {}
